requirements_simple.txt
docx2pdf>=0.1.8
pdf2docx>=0.5.0
python-docx>=0.8.11
reportlab>=3.6.0
Pillow>=8.0.0